Fix BigInt modulo assignment
